requirements.txt
streamlit
pandas
numpy
torch
transformers
langchain
chromadb
sentence-transformers
openai-whisper
pypdf2
groq
python-dotenv
PyMuPDF
pydub
qdrant-client
python-pptx



